# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Schedurx Backend API
  version: 1.0.0
  description: >
    Clinic business API — patients, doctors, appointments, and calendar tools.
    The /tools/* paths are called directly by the Ultravox voice agent during a
    live call; /internal/* is called only by schedurx-ultravox-demo-api.
servers:
  - url: https://backend.example.test
    description: Replace with this service's public URL for deployed environments.
components:
  securitySchemes:
    ToolsBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: TOOLS_API_KEY — configured into Ultravox's tool definitions by the demo bridge.
    InternalBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: INTERNAL_API_KEY — held only by schedurx-ultravox-demo-api.
    FirebaseBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        A Firebase ID token for a Staff account. Custom claims (role, clinicId,
        doctorId) are set at onboarding time via POST /internal/staff and read
        by every /api/v1/* route — clinicId is always sourced from these claims,
        never from the request itself.
    NettuWebhookKeyAuth:
      type: apiKey
      in: header
      name: nettu-scheduler-webhook-key
      description: >
        Static key set once via `PUT /api/v1/account/webhook` on the nettu
        account (see scripts/setup-nettu-webhook.js) — nettu echoes this exact
        value back on every reminder callback. Not a bearer token.
    TwilioSignatureAuth:
      type: apiKey
      in: header
      name: X-Twilio-Signature
      description: >
        HMAC-SHA1 signature over the exact request URL + form params, verified
        against TWILIO_AUTH_TOKEN (see twilio-client.js's validateSignature).
        Twilio computes and sends this automatically; it isn't a credential a
        caller sets manually.
  schemas:
...
    StructuredToolError:
      type: object
      required: [success, error]
      properties:
        success:
          const: false
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type:
                - object
                - "null"
              additionalProperties: true
...
    ApiV1Success:
      type: object
      required: [success, data, message]
      properties:
        success:
          const: true
        data:
          type: object
          additionalProperties: true
        message:
          type:
            - string
            - "null"
    ApiV1Error:
      allOf:
        - $ref: "#/components/schemas/StructuredToolError"
...
paths:
...
  /api/v1/appointments:
    get:
      summary: List appointments for the caller's clinic.
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: date, in: query, schema: { type: string, format: date } }
        - { name: doctorId, in: query, schema: { type: string } }
      responses:
        "200":
          description: Appointments for the caller's clinic only.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
    post:
      summary: Book a new appointment (dashboard reception flow).
      description: >
        Reuses the same appointmentSvc.bookAppointment the voice-agent tool
        uses — nettu-scheduler conflict detection and booking-window rules
        apply identically regardless of who's booking. Resolves the patient
        via findOrCreatePatient server-side.
      security: [{ FirebaseBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [doctorId, start, patient]
              properties:
                doctorId: { type: string }
                start: { type: string, description: ISO 8601 instant. }
                end: { type: [string, "null"] }
                reason: { type: [string, "null"] }
                notes: { type: [string, "null"] }
                bookerRelation: { type: string, enum: [self, proxy] }
                proxyName: { type: [string, "null"] }
                patient:
                  type: object
                  required: [phone]
                  properties:
                    phone: { type: string }
                    name: { type: [string, "null"] }
      responses:
        "201":
          description: Appointment booked.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
        "422":
          description: Missing fields, slot outside booking window, or scheduling rule violation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }
        "409":
          description: Slot no longer available.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }
...
  /api/v1/notifications:
    get:
      summary: List notifications for the caller (own + clinic-broadcast).
      security: [{ FirebaseBearerAuth: [] }]
      responses:
        "200":
          description: Notifications.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
  /api/v1/notifications/read-all:
    patch:
      summary: Mark all of the caller's notifications read.
      security: [{ FirebaseBearerAuth: [] }]
      responses:
        "200":
          description: Done.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
...
  /api/v1/appointments/block:
    post:
      summary: Reserve calendar time with no patient attached (a real nettu busy event, not just a UI flag).
      security: [{ FirebaseBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [doctorId, start]
              properties:
                doctorId: { type: string }
                start: { type: string }
                end: { type: [string, "null"] }
                reason: { type: [string, "null"] }
      responses:
        "201":
          description: Block created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
        "409":
          description: A real booking already exists in that window.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }
  /api/v1/appointments/{id}:
    patch:
      summary: Reschedule an appointment.
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [doctorId, newStart]
              properties:
                doctorId: { type: string }
                newStart: { type: string }
                newEnd: { type: [string, "null"] }
                reason: { type: [string, "null"] }
      responses:
        "200":
          description: Rescheduled — status stays "booked" (or "blocked"), never a literal "rescheduled".
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
        "422":
          description: Outside the reschedule cutoff window, or missing fields.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }
        "409":
          description: New slot no longer available.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }
    delete:
      summary: Cancel an appointment.
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties: { reason: { type: [string, "null"] } }
      responses:
        "200":
          description: Cancelled.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
        "422":
          description: Outside the cancellation cutoff window, or already cancelled.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }

  # ── /api/v1/doctors ────────────────────────────────────────────────────────────
...
  /api/v1/notifications/{id}/read:
    patch:
      summary: Mark one notification read.
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Updated notification.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
  /api/v1/notifications/{id}:
    delete:
      summary: Remove a notification.
      security: [{ FirebaseBearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        "204":
          description: Deleted.

  # ── /api/v1/analytics (utilization, practice-pulse) ───────────────────────────
...
  /api/v1/public/appointments:
    post:
      summary: Book an appointment from the public patient-facing form.
      description: >
        Resolves the patient via findOrCreatePatient server-side — never
        trusts a client-supplied patientId. Fires the booking_confirmed
        WhatsApp workflow, same as every other booking path.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clinicId, doctorId, start, patient]
              properties:
                clinicId: { type: string }
                doctorId: { type: string }
                start: { type: string }
                end: { type: [string, "null"] }
                patient:
                  type: object
                  required: [phone]
                  properties:
                    phone: { type: string, description: "10-digit Indian mobile, optionally +91/91/0-prefixed." }
                    fullName: { type: [string, "null"] }
                    age: { type: [integer, "null"] }
                    gender: { type: [string, "null"] }
                reason: { type: [string, "null"] }
                notes: { type: [string, "null"] }
                bookerRelation: { type: [string, "null"], enum: [self, proxy, null] }
                proxyName: { type: [string, "null"] }
      responses:
        "201":
          description: "{ appointment, patient }"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
        "422":
          description: Missing fields, invalid phone, or scheduling rule violation.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }
        "409":
          description: Slot no longer available.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }
  /api/v1/public/appointments/{id}:
    get:
      summary: Appointment + patient + doctor summary, for the confirmation/manage page.
      description: Requires clinicId to match — the unguessable-link capability model, not a separate auth check.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: clinicId, in: query, required: true, schema: { type: string } }
      responses:
        "200":
          description: "{ appointment, patient, doctor, clinic }"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
        "404":
          description: Not found (wrong clinicId also 404s, not 403 — doesn't confirm the id exists elsewhere).
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }
    patch:
      summary: Reschedule an appointment from the public manage page.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clinicId, newStart]
              properties:
                clinicId: { type: string }
                doctorId: { type: [string, "null"], description: Defaults to the appointment's existing doctor. }
                newStart: { type: string }
                newEnd: { type: [string, "null"] }
                reason: { type: [string, "null"] }
      responses:
        "200":
          description: Rescheduled.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
        "422":
          description: Outside the reschedule cutoff window.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }
    delete:
      summary: Cancel an appointment from the public manage page.
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clinicId]
              properties:
                clinicId: { type: string }
                reason: { type: [string, "null"] }
      responses:
        "200":
          description: Cancelled.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Success" }
        "422":
          description: Outside the cancellation cutoff window, or already cancelled.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiV1Error" }

  # ── /internal/staff/invites — bootstrap only ──────────────────────────────────
...